refactor(ec2): Change the check from ec2 service to vpc service
