+++ prowler/providers/aws/services/vpc/vpc_service.py
@@ -31,6 +31,8 @@ class VPC(AWSService):
         self.vpc_subnets = {}
         self.__threading_call__(self._describe_vpc_subnets)
         self._describe_network_interfaces()
+        self.vpn_connections = {}
+        self.__threading_call__(self._describe_vpn_connections)
 
     def _describe_vpcs(self, regional_client):
         logger.info("VPC - Describing VPCs...")
@@ -398,6 +400,35 @@ class VPC(AWSService):
                 f"{regional_client.region} -- {error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
             )
 
+    def _describe_vpn_connections(self, regional_client):
+        try:
+            describe_vpn_connections = regional_client.describe_vpn_connections()
+
+            for vpn_connection in describe_vpn_connections["VpnConnections"]:
+                arn = f"arn:{self.audited_partition}:ec2:{regional_client.region}:{self.audited_account}:vpn-connection/{vpn_connection['VpnConnectionId']}"
+                if not self.audit_resources or (
+                    is_resource_filtered(arn, self.audit_resources)
+                ):
+                    tunnels = []
+                    for tunnel in vpn_connection["VgwTelemetry"]:
+                        tunnels.append(
+                            VpnTunnel(
+                                status=tunnel["Status"],
+                                outside_ip_address=tunnel["OutsideIpAddress"],
+                            )
+                        )
+                    self.vpn_connections[arn] = VpnConnection(
+                        id=vpn_connection["VpnConnectionId"],
+                        tunnels=tunnels,
+                        region=regional_client.region,
+                        tags=vpn_connection.get("Tags"),
+                    )
+
+        except Exception as error:
+            logger.error(
+                f"{regional_client.region} -- {error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
+            )
+
 
 class VpcSubnet(BaseModel):
     arn: str
@@ -463,5 +494,17 @@ class VpcEndpointService(BaseModel):
     service: str
     owner_id: str
     allowed_principals: list = []
+    region: str
+    tags: Optional[list] = []
+
+
+class VpnTunnel(BaseModel):
+    status: str
+    outside_ip_address: str
+
+
+class VpnConnection(BaseModel):
+    id: str
+    tunnels: list[VpnTunnel]
     region: str
     tags: Optional[list] = []
